Protect about failure to reopen map immediately after it is closed
Remove RandomAccessFiles closed by interrupt
Merge file locks into canonical files store to avoid desync

Fixes ChronicleEnterprise/Chronicle-Map-Enterprise#68
